Move DynamicPriceProvider to com.pablo67340.guishop.api package
